+++ examples/resources/redash_user/import.sh
@@ -1,0 +1,1 @@
+terraform import redash_user.my_user 111 # User ID