+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 ## [Unreleased]
 
+* N/A
+
+## [0.2.0] - 2023-04-25
+
 ### Added
 
+* Add `redash_user` resource.
 * Add debug output.
+
+### Changed
+
+* Ignore disabled users.
 
 ## [0.1.0] - 2023-04-25